+++ src/theme.d.ts
@@ -1,0 +1,5 @@
+import type { Theme } from './index.d.ts';
+
+declare const theme: Theme;
+
+export default theme;